pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wa_toolkit"
version = "0.1.0"
description = "Reusable framework for multi-turn WhatsApp chatbots"
authors = [
    { name = "Replate Engineering" }
]
dependencies = [
    "supabase>=2.11.0",
    "google-genai>=1.2.0",
    "tenacity>=9.0.0",
    "pydantic>=2.0.0"
]
requires-python = ">=3.11"

[tool.setuptools.packages.find]
where = ["."]
include = ["wa_toolkit*"]
